cmsc508-fa2023-prj-intramural-db/pyproject.toml
[tool.poetry]
name = "database project"
version = "0.1.0"
description = "Final project"
authors = ["Stephen Nocera, Tyrese Whitehead"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
tabulate = "^0.9.0"
sqlalchemy = "^2.0.23"
ipython = "^8.18.1"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
